Cambios en Rol y Permiso/Comp   - ObtenerRol (BLL) devuelve el rol solicitado por nombre con una lista cargada con todos los permisos que tenga asociados (simples y compuestos)   - RolDAL tiene un nuevo metodo que instancia y guarda en una lista todos los permisos que estan asociados con el rol que se quiera cargar con datos   - Cambios en PermisoDAL para poder trabajar con permisos simples y compuestos

+++ DAL/Usuario/PermisoDAL.vb
@@ -4,26 +4,54 @@ Imports System.Data.SqlClient
 Public Class PermisoDAL
     Private Shared mProximoID As Integer
 
-    'Ejecuta una query sobre la base para saber cual es el ultimo ID de la tabla y le suma uno
+    ''' <summary>
+    '''Ejecuta una query sobre la base para saber cual es el ultimo ID de la tabla y le suma uno 
+    ''' </summary>
+    ''' <returns>ID a asignar</returns>
+    <Obsolete("El ID deberia manejarse en la BD")>
     Public Shared Function GetProximoID() As Integer
         Return BD.ExecuteScalar("select isnull(max(Permiso_id), 0) from Permiso") + 1
     End Function
 
 
-    'Carga un objeto BE con datos tomados de una fila de la tabla BD
-    Private Shared Function CargarBE(pPermiso As PermisoBE, pRow As DataRow) As PermisoBE
-        pPermiso.ID = pRow("Permiso_id")
-        pPermiso.Nombre = pRow("Permiso_nombre")
-        pPermiso.Padre = pRow("Permiso_padre")
+    ''' <summary>
+    ''' Carga un objeto BE con datos tomados de una fila de la tabla BD
+    ''' </summary>
+    ''' <param name="pPermiso">Objeto BE a Cargar</param>
+    ''' <param name="pRow">Fila con datos a cargar en pPermiso</param>
+    ''' <returns>Objeto BE Con propiedades cargadas</returns>
+    Friend Shared Function CargarBE(pPermiso As PermisoAbstractoBE, pRow As DataRow) As PermisoAbstractoBE
+
+        If TypeOf (pPermiso) Is PermisoBE Then
+            pPermiso.ID = pRow("Permiso_id")
+            pPermiso.Nombre = pRow("Permiso_nombre")
+            pPermiso.Padre = pRow("Permiso_padre")
+        ElseIf TypeOf (pPermiso) Is PermisoCompuestoBE Then
+            pPermiso.ID = pRow("permisoCompuesto_id")
+            pPermiso.Nombre = pRow("permisoCompuesto_nombre")
+            pPermiso.Padre = pRow("permisoCompuesto_padre")
+        End If
 
         Return pPermiso
     End Function
 
 
-    'Ejecuta un query que obtiene los datos de una Permiso
-    Public Shared Function ObtenerPermiso(pID As Integer) As PermisoBE
-        Dim mPermiso As New PermisoBE
-        Dim mCommand As String = "SELECT Permiso_id, Permiso_nombre, permiso_padre FROM Permiso WHERE Permiso_id = " & pID
+    ''' <summary>
+    ''' Ejecuta un query que obtiene los datos de una Permiso
+    ''' </summary>
+    ''' <param name="pID">ID del permiso que se quiere obtener</param>
+    ''' <returns>Datos del permiso indicado recuperados de BD</returns>
+    Public Shared Function ObtenerPermiso(pID As Integer, pCompuesto As Boolean) As PermisoAbstractoBE
+        Dim mPermiso As PermisoAbstractoBE
+        Dim mCommand As String
+
+        If pCompuesto Then
+            mPermiso = New PermisoCompuestoBE
+            mCommand = "SELECT PermisoCompuesto_id, PermisoCompuesto_nombre, permisoCompuesto_padre FROM PermisoCompuesto WHERE PermisoCompuesto_id = " & pID
+        Else
+            mPermiso = New PermisoBE
+            mCommand = "SELECT Permiso_id, Permiso_nombre, permiso_padre FROM Permiso WHERE Permiso_id = " & pID
+        End If
 
         Try
             Dim mDataSet As DataSet = BD.ExecuteDataSet(mCommand)
@@ -42,9 +70,20 @@ Public Class PermisoDAL
     End Function
 
 
-    'Crea un nuevo registro en la tabla Permiso
+    ''' <summary>
+    ''' Crea un nuevo registro en la tabla Permiso
+    ''' </summary>
+    ''' <param name="pPermiso">Objeto BE con datos a persistir en BD</param>
     Public Shared Sub GuardarNuevo(pPermiso As PermisoAbstractoBE)
-        Dim mCommand As String = "INSERT INTO Permiso(Permiso_id, Permiso_nombre, permiso_padre) VALUES (" & pPermiso.ID & ", '" & pPermiso.Nombre & "', " & pPermiso.Padre & ");"
+        Dim mCommand As String = ""
+
+        If TypeOf (pPermiso) Is PermisoBE Then
+            mCommand = "INSERT INTO Permiso(Permiso_id, Permiso_nombre, permiso_padre)
+                        VALUES (" & pPermiso.ID & ", '" & pPermiso.Nombre & "', " & pPermiso.Padre & ");"
+        ElseIf TypeOf (pPermiso) Is PermisoCompuestoBE Then
+            mCommand = "INSERT INTO Permiso(PermisoCompuesto_id, PermisoCompuesto_nombre, permisoCompuesto_padre)
+                        VALUES (" & pPermiso.ID & ", '" & pPermiso.Nombre & "', " & pPermiso.Padre & ");"
+        End If
 
         Try
             BD.ExecuteNonQuery(mCommand)
@@ -55,12 +94,24 @@ Public Class PermisoDAL
     End Sub
 
 
-    'Modifica un registro de la tabla Permiso
+    ''' <summary>
+    ''' Modifica un registro de la tabla Permiso
+    ''' </summary>
+    ''' <param name="pPermiso">Objeto BE con datos modificados a persistir</param>
     Public Shared Sub GuardarModificacion(pPermiso As PermisoAbstractoBE)
-        Dim mCommand As String = "UPDATE Permiso SET " &
+        Dim mCommand As String = ""
+
+        If TypeOf (pPermiso) Is PermisoBE Then
+            mCommand = "UPDATE Permiso SET " &
                                  "Permiso_nombre = '" & pPermiso.Nombre &
                                  "', permiso_padre = " & pPermiso.Padre &
                                  " WHERE Permiso_id = " & pPermiso.ID
+        ElseIf TypeOf (pPermiso) Is PermisoCompuestoBE Then
+            mCommand = "UPDATE Permiso SET " &
+                                 "PermisoCompuesto_nombre = '" & pPermiso.Nombre &
+                                 "', permisoCompuesto_padre = " & pPermiso.Padre &
+                                 " WHERE PermisoCompuesto_id = " & pPermiso.ID
+        End If
 
         Try
             BD.ExecuteNonQuery(mCommand)
@@ -70,10 +121,19 @@ Public Class PermisoDAL
         End Try
     End Sub
 
-
-    'Elimina un registro de la tabla Permiso
+    'Agregar implementacion para eliminar relaciones
+    ''' <summary>
+    ''' Elimina un registro de la tabla Permiso
+    ''' </summary>
+    ''' <param name="pPermiso">Objeto BE con datos a eliminar en BD</param>
     Public Shared Sub Eliminar(pPermiso As PermisoAbstractoBE)
-        Dim mCommand As String = "DELETE FROM Permiso WHERE Permiso_id = " & pPermiso.ID
+        Dim mCommand As String = ""
+
+        If TypeOf (pPermiso) Is PermisoBE Then
+            mCommand = "DELETE FROM Permiso WHERE Permiso_id = " & pPermiso.ID
+        ElseIf TypeOf (pPermiso) Is PermisoCompuestoBE Then
+            mCommand = "DELETE FROM PermisoCompuesto WHERE PermisoCompuesto_id = " & pPermiso.ID
+        End If
 
         Try
             BD.ExecuteNonQuery(mCommand)
@@ -84,10 +144,20 @@ Public Class PermisoDAL
     End Sub
 
 
-    'Devuelve una lista de objetos PermisoBE con los datos de cada registro de la tabla Permiso
-    Public Shared Function ListarPermisos() As List(Of PermisoAbstractoBE)
+    ''' <summary>
+    ''' Devuelve una lista de objetos PermisoBE con los datos de cada registro de la tabla Permiso o PermisoCompuesto
+    ''' </summary>
+    ''' <returns>List(Of PermisoAbstracto) con todos los permisos simples o compuestos existentes en BD</returns>
+    Public Shared Function ListarPermisos(pCompuesto As Boolean) As List(Of PermisoAbstractoBE)
         Dim mLista As New List(Of PermisoAbstractoBE)
-        Dim mCommand As String = "SELECT Permiso_id, Permiso_nombre, permiso_padre FROM Permiso"
+        Dim mCommand As String = ""
+
+        If pCompuesto Then
+            mCommand = "SELECT PermisoCompuesto_id, PermisoCompuesto_nombre, permisoCompuesto_padre FROM PermisoCompuesto"
+        Else
+            mCommand = "SELECT Permiso_id, Permiso_nombre, permiso_padre FROM Permiso"
+        End If
+
         Dim mDataSet As DataSet
 
         Try
@@ -95,10 +165,15 @@ Public Class PermisoDAL
 
             If Not IsNothing(mDataSet) And mDataSet.Tables.Count > 0 And mDataSet.Tables(0).Rows.Count > 0 Then
                 For Each mRow As DataRow In mDataSet.Tables(0).Rows
-                    Dim mPermiso As New PermisoBE
+                    Dim mPermiso As PermisoAbstractoBE
+
+                    If pCompuesto Then
+                        mPermiso = New PermisoCompuestoBE
+                    Else
+                        mPermiso = New PermisoBE
+                    End If
 
                     mPermiso = CargarBE(mPermiso, mRow)
-
                     mLista.Add(mPermiso)
                 Next
